# By Markus Döring (3) and others
# Via Kyle Braak
* 'master' of https://github.com/gbif/occurrence:
  even more tests added to test dataset
  add verbatim only migration script; deprecate non-verbatim version
  Added bare bones parsers for missing occurrence enums: sex, establishment means, lifestage
  Fix yet another double/int broken elevation test
  Moved InternalTerm to dwc-api

+++ occurrence-ws-client/src/test/java/org/gbif/occurrence/ws/client/mock/OccurrencePersistenceMockServiceTest.java
@@ -22,6 +22,6 @@ public class OccurrencePersistenceMockServiceTest {
     assertEquals(BasisOfRecord.PRESERVED_SPECIMEN, occ.getBasisOfRecord());
     assertEquals((Integer) 1912, occ.getYear());
     assertEquals((Integer) 5, occ.getMonth());
-    assertEquals((Integer) 112, occ.getElevation());
+    assertEquals( (Double) 112d, occ.getElevation());
   }
 }
